Add workshop check for error code 134
There seemingly is an assert that checks if the entirety of a workshop item downloaded. This will cause it to exit with error code 134 on large downloads, where it only succeeds in downloading a part of it at a time before timing out. Repeating it multiple times allows it to eventually finish, the amount of times depending on the size of the download.

This pull request makes it check for SteamCMD error 134, and retry if the validate flag is set, the same way return code 10 is handled.